requirements.txt
vonage
python-dotenv
'fastapi[standard]'